refactor: extract shared StatCard widget from game_screen and hud_overlay
Both screens had near-identical private _StatCard widgets with trivially
diverging color values. Consolidate into a single public StatCard with an
explicit accentColor parameter.
